- completed addition of lazy loader - updated ReadME with Lazy Loader reference
