+++ README.md
@@ -11,3 +11,6 @@ I am currently in the mode of slowly refactoring the CSS but I think I may event
 #### Resume API
 Much of my resume data is coming from https://gitconnected.com where you can create your own resume API for FREE!
 SUPPORT this site :-)
+
+#### Lazy loader reference
+https://medium.com/the-non-traditional-developer/how-to-use-an-intersectionobserver-in-a-react-hook-9fb061ac6cb5
