+++ src/components/AboutComponent/about.component.jsx
@@ -5,9 +5,8 @@ import blackManLearning from '../../images/black-man-learning.png';
 import blackManGraduate from '../../images/black-man-graduate.png';
 import blackManReading from '../../images/black-man-reading.png';
 import blackPeopleDancing from '../../images/black-people-dancing.png';
-import {
-  staggered, unstaggered, container, aboutImg,
-} from './about.module.scss';
+import { staggered, unstaggered, container } from './about.module.scss';
+import ImageLazyLoader from './ImageLazyLoader';
 
 const AboutComponent = () => (
   <div>
@@ -25,11 +24,7 @@ const AboutComponent = () => (
                   <h2 className="colorlib-heading">Who Am I?</h2>
                   <div className={staggered}>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackFamily}
-                        alt="black family sitting on couch"
-                      />
+                      <ImageLazyLoader src={blackFamily} width={null} height="200px" alt="black family sitting on couch" />
                       <p>
                         I am loving husband.
                         <br />
@@ -38,34 +33,18 @@ const AboutComponent = () => (
                     </div>
                     <div className={container}>
                       <p>I am a computer science graduate of Virginia Tech, class of 2007</p>
-                      <img
-                        className={aboutImg}
-                        src={blackManGraduate}
-                        alt="black man graduate"
-                      />
+                      <ImageLazyLoader src={blackManGraduate} width={null} height="200px" alt="black man graduate" />
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackManUsingLaptop}
-                        alt="black man programming"
-                      />
+                      <ImageLazyLoader src={blackManUsingLaptop} width={null} height="200px" alt="black man programming" />
                       <p>I am software and web developer with over 15 years of experience.</p>
                     </div>
                     <div className={container}>
                       <p>I have a passion to continuously learn.</p>
-                      <img
-                        className={aboutImg}
-                        src={blackManLearning}
-                        alt="black man write on white board"
-                      />
+                      <ImageLazyLoader src={blackManLearning} width={null} height="200px" alt="black man write on white board" />
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackManReading}
-                        alt="black man reading on couch"
-                      />
+                      <ImageLazyLoader src={blackManReading} width={null} height="200px" alt="black man reading on couch" />
                       {/* eslint-disable-next-line max-len */}
                       <p>
                         {`I love reading all types of books but when it comes to web development
@@ -74,20 +53,14 @@ const AboutComponent = () => (
                     </div>
                     <div className={container}>
                       <p>I am a curator of music; a DJ and hip hop enthusiast.</p>
-                      <img
-                        className={aboutImg}
-                        src={blackPeopleDancing}
-                        alt="black people dancing"
-                      />
+                      {/* eslint-disable-next-line max-len */}
+                      <ImageLazyLoader src={blackPeopleDancing} width={null} height="200px" alt="black people dancing" />
                     </div>
                   </div>
                   <div className={unstaggered}>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackFamily}
-                        alt="black family sitting on couch"
-                      />
+                      {/* eslint-disable-next-line max-len */}
+                      <ImageLazyLoader src={blackFamily} width={null} height="200px" alt="black family sitting on couch" />
                       <p>
                         I am loving husband.
                         <br />
@@ -95,46 +68,26 @@ const AboutComponent = () => (
                       </p>
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackManGraduate}
-                        alt="black man graduate"
-                      />
+                      <ImageLazyLoader src={blackManGraduate} width={null} height="200px" alt="black man graduate" />
                       <p>I am a computer science graduate of Virginia Tech, class of 2007</p>
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackManUsingLaptop}
-                        alt="black man programming"
-                      />
+                      <ImageLazyLoader src={blackManUsingLaptop} width={null} height="200px" alt="black man programming" />
                       <p>I am software and web developer with over 15 years of experience.</p>
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackManLearning}
-                        alt="black man write on white board"
-                      />
+                      <ImageLazyLoader src={blackManLearning} width={null} height="200px" alt="black man write on white board" />
                       <p>I have a passion to continuously learn.</p>
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackManReading}
-                        alt="black man reading on couch"
-                      />
+                      <ImageLazyLoader src={blackManReading} width={null} height="200px" alt="black man reading on couch" />
                       <p>
                         {`I love reading all types of books but when it comes to web development
                         I like book series such as "You Dont Know JS" by Kyle Simpson.`}
                       </p>
                     </div>
                     <div className={container}>
-                      <img
-                        className={aboutImg}
-                        src={blackPeopleDancing}
-                        alt="black people dancing"
-                      />
+                      <ImageLazyLoader src={blackPeopleDancing} width={null} height="200px" alt="black people dancing" />
                       <p>I am a curator of music; a DJ and hip hop enthusiast.</p>
                     </div>
                   </div>
