+++ src/index.js
@@ -1,10 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
-import { createBrowserHistory } from "history";
-import { HashRouter } from "react-router-dom";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import { createBrowserHistory } from 'history';
+import { HashRouter } from 'react-router-dom';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
 
 const history = createBrowserHistory();
 
@@ -12,7 +12,7 @@ ReactDOM.render(
   <HashRouter history={history}>
     <App />
   </HashRouter>,
-  document.getElementById("root")
+  document.getElementById('root'),
 );
 
 // If you want your app to work offline and load faster, you can change
